% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aof.R
\name{aof}
\alias{aof}
\title{aof}
\usage{
aof(bee, Age, x)
}
\arguments{
\item{bee}{The identity of the bee as factor (e.g. "A00103C00020C301", "bee1").}

\item{Age}{The age of the bee in day as numeric (e.g. 1, 4, 32).}

\item{x}{The daily activity of the bee at a given age as a numeric value,
for instance (i) the number of the trips per day, (ii) the duration of
the trips per day, or (iii) the time of the trips per day.}
}
\value{
A data.frame containing the aof results (one row per bee).
}
\description{
Ontogenetic shifts in central-place foraging insects
}
\details{
A breakpoint-based method to detect ontogenetic shifts in
univariate time-activity budget series of central-place foraging insects.
The method finds a single breakpoint according to the likelihood function.
The method was developped with honey bees in order to detect the
Age at Onset of Foraging (AOF), but can be used for the detection of other
ontongenetic shifts in other central-place foraging insects. For more
details, see Requier et al. (2020) Measuring ontogenetic shifts in
central-place foraging insects. Journal of Animal Ecology.
}
\examples{
require("bcpa")
# Exemple with simulated data:
# No change simulated
mu1 <- 50  # behavioral change: 25 or 50
mu2 <- 50
rho1 = 0.5 # intrerval frequency
rho2 = 0.5
# N low and V low
# A single changepoint in a time-series where the parameters change at
# some unknown timepoints t* is done by simply sweeping all possible
# breaks, and finding the most-likely changepoint according to the
# likelihood. This is illustrated below:
n.obs <- 5 # no. observations: 5 to 45
sigma1 <- 0.1 # variance: 0.1 to 3
sigma2 <- sigma1
SimTS <- function(n, mu, rho, sigma){
  X.standard <- arima.sim(n, model = list(ar = rho))
  X.standard/sd(X.standard)*sigma + mu
}
# create time series with break at 25
t.full <- 0:50
t.break <- 25
x.full <- c(SimTS(t.break, mu1, rho1, sigma1),
  SimTS(max(t.full)-t.break+1, mu2, rho2, sigma2))
# subsample of observations (n defined above) and estimate
keep <- sort(sample(1:length(x.full), n.obs))
TimeBudget <- data.frame(
  bee = "A",
  Age = t.full[keep],
  x = x.full[keep])
# Running the algorithm
AOF <- aof(
  bee = TimeBudget$bee,
  Age = TimeBudget$Age,
  x = TimeBudget$x)
print(AOF)
# see vignette for more examples
}
