% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataExample}
\alias{dataExample}
\title{subset of 5 bees.}
\format{A data.frame with three columns.
\describe{
  \item{name}{The identity of the bee}
  \item{Age}{The age of the bee in day.}
  \item{Number}{Number of trips per day.}
  \item{Duration}{duration of trips per day.}
  \item{Time}{Time of trips per day.}
}}
\usage{
dataExample
}
\description{
This is a subset of 5 bees randomly selected in the
  exeprimental design of Requier et al. (J. Animal Ecology).
}
\keyword{datasets}
