% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osfi.R
\name{osfi}
\alias{osfi}
\title{osfi}
\usage{
osfi(bee, Age, x, getPlot = FALSE)
}
\arguments{
\item{bee}{xxxx}

\item{Age}{xxx}

\item{x}{xxxx}

\item{getPlot}{A boolean for graphical output.}
}
\value{
xxxx
}
\description{
osfi
}
\details{
xxxx
}
\examples{
require("bcpa")
par(mar = c(4, 4, 1, 1), mfrow = c(1, 1))
# 1. Exemple with simulated data:
# 1.1. No change simulated
mu1 <- 50  # behavioral change: 25 or 50
mu2 <- 50
rho1 = 0.5 # intrerval frequency
rho2 = 0.5
# 1.1.1. N low and V low
# A single changepoint in a time-series where the parameters change at
# some unknown timepoints t* is done by simply sweeping all possible
# breaks, and finding the most-likely changepoint according to the
# likelihood. This is illustrated below:
n.obs <- 5 # no. observations: 5 to 45
sigma1 <- 0.1 # variance: 0.1 to 3
sigma2 <- sigma1
SimTS <- function(n, mu, rho, sigma){
  X.standard <- arima.sim(n, model = list(ar = rho))
  X.standard/sd(X.standard)*sigma + mu
}
# create time series with break at 500
t.full <- 0:50
t.break <- 25
x.full <- c(SimTS(t.break, mu1, rho1, sigma1),
  SimTS(max(t.full)-t.break+1, mu2, rho2, sigma2))
# subsample 100 observations and estimate
keep <- sort(sample(1:length(x.full), n.obs))
TimeBudget <- data.frame(
  bee = "A",
  Age = t.full[keep],
  x = x.full[keep])
# Running the algorithm
AOF <- osfi(
  bee = TimeBudget$bee,
  Age = TimeBudget$Age,
  x = TimeBudget$x,
  getPlot = TRUE)
print(AOF)



# 1.1.2.
}
